+++ lib/src/pubspec.yaml.g.dart
@@ -92,13 +92,13 @@ sealed class Pubspec {
   static const PubspecVersion version = (
     /// Non-canonical string representation of the version as provided
     /// in the pubspec.yaml file.
-    representation: r'1.2.0',
+    representation: r'1.2.3',
 
     /// Returns a 'canonicalized' representation
     /// of the application version.
     /// This represents the version string in accordance with
     /// Semantic Versioning (SemVer) standards.
-    canonical: r'1.2.0',
+    canonical: r'1.2.3',
 
     /// MAJOR version when you make incompatible API changes.
     /// The major version number: 1 in "1.2.3".
@@ -111,7 +111,7 @@ sealed class Pubspec {
 
     /// PATCH version when you make backward compatible bug fixes.
     /// The patch version number: 3 in "1.2.3".
-    patch: 0,
+    patch: 3,
 
     /// The pre-release identifier: "foo" in "1.2.3-foo".
     preRelease: <String>[],
@@ -123,13 +123,13 @@ sealed class Pubspec {
   /// Build date and time (UTC)
   static final DateTime timestamp = DateTime.utc(
     2026,
-    4,
-    13,
-    20,
-    38,
-    6,
-    354,
-    898,
+    5,
+    26,
+    23,
+    8,
+    40,
+    314,
+    946,
   );
 
   /// Name
@@ -165,7 +165,7 @@ sealed class Pubspec {
   /// Users see it when they [browse for packages](https://pub.dev/packages).
   /// The description is plain text: no markdown or HTML.
   static const String description =
-      r'An easy to use cross-platform regex replace command line util.';
+      r'An easy to use cross-platform regex replace command line util and pm CLI to modify pubspec.yaml files.';
 
   /// Homepage
   ///